# Use WMI to implement Volume API to reduce PowerShell overhead