--- a/go.mod
+++ b/go.mod
@@ -9,10 +9,12 @@ toolchain go1.22.3
 
 require (
 	github.com/Microsoft/go-winio v0.6.2
+	github.com/go-ole/go-ole v1.3.0
 	github.com/google/go-cmp v0.6.0
 	github.com/grpc-ecosystem/go-grpc-middleware/providers/prometheus v1.0.1
 	github.com/iancoleman/strcase v0.3.0
 	github.com/kubernetes-csi/csi-proxy/client v1.1.3
+	github.com/microsoft/wmi v0.25.1
 	github.com/pkg/errors v0.9.1
 	github.com/prometheus/client_golang v1.20.5
 	github.com/sergi/go-diff v1.3.1