… go.sum
… pkg/os/volume/api.go
… vendor/github.com/go-ole/go-ole/.travis.yml
… vendor/github.com/go-ole/go-ole/ChangeLog.md
… vendor/github.com/go-ole/go-ole/LICENSE
… vendor/github.com/go-ole/go-ole/README.md
… vendor/github.com/go-ole/go-ole/SECURITY.md
… vendor/github.com/go-ole/go-ole/appveyor.yml
… vendor/github.com/go-ole/go-ole/com.go
… vendor/github.com/go-ole/go-ole/com_func.go
… vendor/github.com/go-ole/go-ole/connect.go
… vendor/github.com/go-ole/go-ole/constants.go
… vendor/github.com/go-ole/go-ole/error.go
… vendor/github.com/go-ole/go-ole/error_func.go
… vendor/github.com/go-ole/go-ole/error_windows.go
… vendor/github.com/go-ole/go-ole/guid.go
… vendor/github.com/go-ole/go-ole/iconnectionpoint.go
… vendor/github.com/go-ole/go-ole/iconnectionpoint_func.go
… vendor/github.com/go-ole/go-ole/iconnectionpoint_windows.go
… vendor/github.com/go-ole/go-ole/iconnectionpointcontainer.go
… vendor/github.com/go-ole/go-ole/iconnectionpointcontainer_func.go
… vendor/github.com/go-ole/go-ole/iconnectionpointcontainer_windows.go
… vendor/github.com/go-ole/go-ole/idispatch.go
… vendor/github.com/go-ole/go-ole/idispatch_func.go
… vendor/github.com/go-ole/go-ole/idispatch_windows.go
… vendor/github.com/go-ole/go-ole/ienumvariant.go
… vendor/github.com/go-ole/go-ole/ienumvariant_func.go
… vendor/github.com/go-ole/go-ole/ienumvariant_windows.go
… vendor/github.com/go-ole/go-ole/iinspectable.go
… vendor/github.com/go-ole/go-ole/iinspectable_func.go