@andyzhangx 2024-10-22
thanks for the PR. @laozc have you tested the WMI based operation manually? does it work?