@laozc 2024-10-23
Yes. I verified it in my local environment and they work the same as PowerShell commands.
I'm still working on finalizing the PR with more testing and integrate it into CSI driver to get some result.