@laozc 2025-02-21
Ran some test with two Windows nodes setup, one with the CSI-proxy with PowerShell, while another one with the CSI-proxy using WMI.
The test creates two pods side by side and delete the resources from the cluster once they're running.

* Kubernetes: version 1.31.4
* CSI: vSphere CSI
* CNI: Antrea
* Windows Node Spec: 4 vCPU, 16GB memory

Here is the performance data with collected from the gRPC server metrics within 10 minutes time frame.

| Operation | Before (sec) | After (sec) | Before (Min) | After (Max) | Improvement |
|-----------|--------------|-------------|--------------|-------------|-------------|
| FormatVolume (1G disk) | 5.85-8.85 | 2.90s | 5.85 | 2.90 | -101% |
| GetDiskNumberFromVolumeID | 2.90 | 0.29-0.58 | 2.90 | 0.58 | -400% |
| GetVolumeIDFromTargetPath | 0.58-0.96 | 0.009 | 0.58 | 0.009 | -6344% |
| GetVolumeStats | 2.90 | 0.29 | 2.90 | 0.29 | -900% |
| IsVolumeFormatted | 2.90 | 0.0955 | 2.90 | 0.0955 | -2937% |
| ListDiskIDs | 2.9-5.85 | 0.585 | 2.90 | 0.585 | -396% |
| ListVolumeOnDisk | 2.9 | 0.29-0.585 | 2.90 | 0.585 | -396% |
| MountVolume | 2.9-5.85 | 0.00955-0.0955 | 2.90 | 0.09 | -3122% |
| PartitionDisk (1G disk) | 19.5 | 2.9 | 19.5 | 2.9 | -572% |
| SetDiskState | 2.9-5.70 | 0.0955-0.280 | 2.90 | 0.280 | -936% |
| UnmountVolume | 5.85-8.85 | 0.29-0.585 | 5.85 | 0.585 | -900% |

It's obviously shown that PowerShell takes around 2.9s to load in the environment while memory consumption could reach up to 200MB.
By reducing the PowerShell overhead, we were able to cut down the time cost significantly to method calls for all operations.