@mauriciopoppe 2025-02-22
One more thing, this PR is against the server version of CSI Proxy, we did lots of development against the library version in https://github.com/kubernetes-csi/csi-proxy/tree/library-development which doesn't use a client/server model anymore, I think it'd be great to see this work in that other branch once there's signal that the PR doesn't cause any regressions.

I see that you tested this against the vSphere CSI Driver, do you know if there are plans for that CSI Driver to use the library version of CSI Proxy?