@laozc 2025-02-24
For the current 1.x release when gRPC is used, I believe we could merge #371 first in case monitoring the gRPC operation performance metrics is preferred, especially we need some concrete evidence for the case like this PR.

I believe vSphere CSI driver is working on moving to CSI proxy library release as well.
I will be working on the same PR for the library version. Once the PR for the library release is ready, we could safely process with both PRs in two branches without causing any regression.