@mauriciopoppe 2025-03-11
Now that https://github.com/kubernetes-csi/csi-proxy/pull/371 let's continue this work! My only concern is that Powershell is way easier to debug than doing syscalls.

Would it be possible to split the PR into multiple PRs grouping by changes per API please? I think it'd be easier to read and review.