@mauriciopoppe 2025-03-12
Let's try on master first, after a new CSI Proxy release it'd be a matter of updating the CSI Proxy binary deployment only, no changes in the CSI Driver.

For the library version, it requires some changes in the CSI Driver.