@laozc 2025-03-13
I believe we could safely remove the constraint to use Go 1.23 with this change.
```
mountedFolder, err := isMountedFolder(candidatePath)
```
This method uses Windows API calls to determine whether the path is a mounted path instead of relying on the Go API behavior.