@mauriciopoppe 2025-04-14
Hey @andyzhangx, would it be possible for you to review the PR please? Maybe you have experience with the `github.com/microsoft/wmi` package and how to use it. From my end, I can only check that the integration tests are still passing.