@andyzhangx 2025-04-15
@laozc I tries to copy the cim folder into my Azure Disk CSI driver repo (for test): https://github.com/kubernetes-sigs/azuredisk-csi-driver/compare/master...andyzhangx:azuredisk-csi-driver:cim?expand=1, but I got following error, have you hit similar error in your CSI driver?

https://github.com/andyzhangx/azuredisk-csi-driver/actions/runs/14463170910/job/40559504163

```
# sigs.k8s.io/azuredisk-csi-driver/pkg/os/cim
FAIL	sigs.k8s.io/azuredisk-csi-driver/pkg/os/cim [setup failed]
package sigs.k8s.io/azuredisk-csi-driver/pkg/os/cim
	imports github.com/microsoft/wmi/pkg/wmiinstance
	imports golang.org/x/sys/windows: build constraints exclude all Go files in /home/runner/work/azuredisk-csi-driver/azuredisk-csi-driver/vendor/golang.org/x/sys/windows
# sigs.k[8](https://github.com/andyzhangx/azuredisk-csi-driver/actions/runs/14463170910/job/40559504163#step:4:9)s.io/azuredisk-csi-driver/pkg/os/cim
package sigs.k8s.io/azuredisk-csi-driver/pkg/os/cim
	imports github.com/microsoft/wmi/server201[9](https://github.com/andyzhangx/azuredisk-csi-driver/actions/runs/14463170910/job/40559504163#step:4:10)/root/microsoft/windows/storage
	imports github.com/microsoft/wmi/pkg/base/instance
	imports github.com/microsoft/wmi/pkg/base/session: build constraints exclude all Go files in /home/runner/work/azuredisk-csi-driver/azuredisk-csi-driver/vendor/github.com/microsoft/wmi/pkg/base/session
```