@laozc 2025-04-15
I could add build constraints to these codes if the downstream projects need.
```
//go:build windows
// +build windows
```